Double declaration has changed

+++ src/main/scala/com/imaginea/Main.scala
@@ -1594,7 +1594,7 @@ object Main extends App {
               case Some(site) =>
                 val listOfInstances = site.instances
                 val listOfInstanceFlavors = listOfInstances.map { instance =>
-                  val zeroUtilization = 0.toDouble
+                  val zeroUtilization = 0.0
                   val memInfo = instance.memoryInfo.map(_.total).getOrElse(zeroUtilization)
                   val dskInfo = instance.rootDiskInfo.map(_.total).getOrElse(zeroUtilization)
                   InstanceFlavor(instance.instanceType.getOrElse(""), None, memInfo, dskInfo)
